style: 형변환 제거, 메소드 위치 변경

+++ src/main/java/lotto/domain/Result.java
@@ -19,10 +19,10 @@ public class Result {
     }
 
     public float calculateRate() {
-        return (float) rankCount.entrySet()
+        return rankCount.entrySet()
                 .stream()
                 .mapToLong(Result::calculateEntryPrice)
-                .sum() / getInputMoney() * 100;
+                .sum() / getInputMoney() * 100.0F;
     }
 
     private static long calculateEntryPrice(final Entry<Rank, Long> entry) {
